+++ python/core/auto_generated/pointcloud/qgspointcloudattribute.sip.in
@@ -9,6 +9,8 @@
 
 
 
+
+
 class QgsPointCloudAttribute
 {
 %Docstring
@@ -75,6 +77,19 @@ Returns ``True`` if the specified data ``type`` is numeric.
 %MethodCode
     QString str = QStringLiteral( "<QgsPointCloudAttribute: %1 (%2)>" ).arg( sipCpp->name() ).arg( sipCpp->displayType() );
     sipRes = PyUnicode_FromString( str.toUtf8().constData() );
+%End
+
+    static void getPointXYZ( const char *ptr, int i, std::size_t pointRecordSize, int xOffset, QgsPointCloudAttribute::DataType xType,
+                             int yOffset, QgsPointCloudAttribute::DataType yType,
+                             int zOffset, QgsPointCloudAttribute::DataType zType,
+                             const QgsVector3D &indexScale, const QgsVector3D &indexOffset, double &x, double &y, double &z );
+%Docstring
+Retrieves the x, y, z values for the point at index ``i``.
+%End
+
+    static QVariantMap getAttributeMap( const char *data, std::size_t recordOffset, const QgsPointCloudAttributeCollection &attributeCollection );
+%Docstring
+Retrieves all the attributes of a point
 %End
 
 };
